Be multi-gamemode aware in listeners.
Some of the listeners were firing in other game modes so I added code to
check for the world and also to prevent accidental use in the future.
i.e., throw and exception. Parkour keeps its own table of islands in the
database and we don't want non-parkour islands to end up in there.

Reworked the CourseRunnerLister test to include the new teleport
listener.